requirements.txt
Tk
pycryptodomex